Backend/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.4.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
dataclasses-json==0.6.7
earthengine-api==1.7.3
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.95.2
fastapi-cache2==0.2.2
fastapi-jwt-auth==0.5.0
filetype==1.2.0
flatbuffers==25.9.23
frozenlist==1.8.0
gast==0.7.0
geographiclib==2.1
geopy==2.4.1
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-cloud-core==2.5.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
greenlet==3.3.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.12.0
langchain==0.0.350
langchain-classic==1.0.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-text-splitters==1.0.0
langgraph==1.0.4
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.2.14
langsmith==0.0.87
libclang==18.1.1
Mako==1.3.10
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.1
mdurl==0.1.2
ml_dtypes==0.5.4
multidict==6.7.0
mypy_extensions==1.1.0
namex==0.1.0
numpy==1.26.4
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.5
ormsgpack==1.12.0
packaging==23.2
pandas==2.3.3
pendulum==3.1.0
pgvector==0.4.2
Pillow==9.5.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==1.10.12
pydantic-settings==2.12.0
Pygments==2.19.2
PyJWT==1.7.1
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
redis==7.1.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rsa==4.9.1
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.44
starlette==0.27.0
tenacity==8.5.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tomli==2.3.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.0
uuid_utils==0.12.0
uvicorn==0.38.0
Werkzeug==3.1.4
wrapt==2.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
